Some refactoring progress
