+++ packages/attendance/src/utils/types/types.ts
@@ -1,0 +1,4 @@
+import { Moment } from "moment";
+// Moment Related
+export type MomentType = Moment;
+export type MomentUnionType = Moment | Moment[] | Moment[][];
